ensure gazelle
