+++ validator/BUILD.bazel
@@ -23,7 +23,6 @@ go_library(
         "//shared/version:go_default_library",
         "//validator/accounts/v1:go_default_library",
         "//validator/accounts/v2:go_default_library",
-        "//validator/client/streaming:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
         "@com_github_joonix_log//:go_default_library",
@@ -65,7 +64,7 @@ go_image(
         "//shared/version:go_default_library",
         "//validator/accounts/v1:go_default_library",
         "//validator/accounts/v2:go_default_library",
-        "//validator/client/streaming:go_default_library",
+        "//validator/client:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
         "@com_github_joonix_log//:go_default_library",
