+++ validator/client/attest.go
@@ -1,7 +1,9 @@
-package streaming
+package client
 
 import (
+	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"time"
 
@@ -16,7 +18,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/roughtime"
 	"github.com/prysmaticlabs/prysm/shared/slotutil"
-	"github.com/prysmaticlabs/prysm/validator/client/metrics"
 	keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v1"
 	"github.com/sirupsen/logrus"
 	"go.opencensus.io/trace"
@@ -33,12 +34,11 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 
 	fmtKey := fmt.Sprintf("%#x", pubKey[:])
 	log := log.WithField("pubKey", fmt.Sprintf("%#x", bytesutil.Trunc(pubKey[:]))).WithField("slot", slot)
-	epoch := slot / params.BeaconConfig().SlotsPerEpoch
-	duty, err := v.duty(pubKey, epoch)
+	duty, err := v.duty(pubKey)
 	if err != nil {
 		log.WithError(err).Error("Could not fetch validator assignment")
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
@@ -57,11 +57,10 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 	if err != nil {
 		log.WithError(err).Error("Could not request attestation to sign at slot")
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
-
 	indexedAtt := &ethpb.IndexedAttestation{
 		AttestingIndices: []uint64{duty.ValidatorIndex},
 		Data:             data,
@@ -69,11 +68,12 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 	if err := v.preSigningValidations(ctx, indexedAtt, pubKey); err != nil {
 		return
 	}
+
 	sig, err := v.signAtt(ctx, pubKey, data)
 	if err != nil {
 		log.WithError(err).Error("Could not sign attestation")
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
@@ -90,7 +90,7 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 	if !found {
 		log.Errorf("Validator ID %d not found in committee of %v", duty.ValidatorIndex, duty.Committee)
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
@@ -107,7 +107,7 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 	if err != nil {
 		log.WithError(err).Error("Could not submit attestation to beacon node")
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
@@ -119,17 +119,16 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 		}).Fatal("made a slashable attestation, found by external slasher service")
 		return
 	}
-
 	if err := v.saveAttesterIndexToData(data, duty.ValidatorIndex); err != nil {
 		log.WithError(err).Error("Could not save validator index for logging")
 		if v.emitAccountMetrics {
-			metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
 		}
 		return
 	}
 
 	if v.emitAccountMetrics {
-		metrics.ValidatorAttestSuccessVec.WithLabelValues(fmtKey).Inc()
+		ValidatorAttestSuccessVec.WithLabelValues(fmtKey).Inc()
 	}
 
 	span.AddAttributes(
@@ -141,6 +140,77 @@ func (v *validator) SubmitAttestation(ctx context.Context, slot uint64, pubKey [
 		trace.Int64Attribute("targetEpoch", int64(data.Target.Epoch)),
 		trace.StringAttribute("bitfield", fmt.Sprintf("%#x", aggregationBitfield)),
 	)
+}
+func (v *validator) preSigningValidations(ctx context.Context, indexedAtt *ethpb.IndexedAttestation, pubKey [48]byte) error {
+	fmtKey := fmt.Sprintf("%#x", pubKey[:])
+	log := log.WithField("pubKey", fmt.Sprintf("%#x", bytesutil.Trunc(pubKey[:]))).WithField("slot", indexedAtt.Data.Slot)
+	if featureconfig.Get().ProtectAttester {
+		v.attesterHistoryByPubKeyLock.RLock()
+		attesterHistory := v.attesterHistoryByPubKey[pubKey]
+		v.attesterHistoryByPubKeyLock.RUnlock()
+		if isNewAttSlashable(attesterHistory, indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch) {
+			log.WithFields(logrus.Fields{
+				"sourceEpoch": indexedAtt.Data.Source.Epoch,
+				"targetEpoch": indexedAtt.Data.Target.Epoch,
+			}).Error("Attempted to make a slashable attestation, rejected")
+			if v.emitAccountMetrics {
+				ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
+			}
+			return fmt.Errorf("sourceEpoch: %dtargetEpoch: %d  Attempted to make a slashable attestation,"+
+				" rejected by local slasher protection", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
+		}
+	}
+	if featureconfig.Get().SlasherProtection && v.protector != nil {
+		if !v.protector.VerifyAttestation(ctx, indexedAtt) {
+			log.WithFields(logrus.Fields{
+				"sourceEpoch": indexedAtt.Data.Source.Epoch,
+				"targetEpoch": indexedAtt.Data.Target.Epoch,
+			}).Error("Attempted to make a slashable attestation, rejected by external slasher service")
+			if v.emitAccountMetrics {
+				ValidatorAttestFailVecSlasher.WithLabelValues(fmtKey).Inc()
+			}
+			return fmt.Errorf("sourceEpoch: %dtargetEpoch: %d  Attempted to make a slashable attestation,"+
+				" rejected by external slasher service", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
+		}
+	}
+	return nil
+}
+
+func (v *validator) postSignatureUpdate(ctx context.Context, indexedAtt *ethpb.IndexedAttestation, pubKey [48]byte) error {
+	fmtKey := fmt.Sprintf("%#x", pubKey[:])
+	if featureconfig.Get().ProtectAttester {
+		v.attesterHistoryByPubKeyLock.Lock()
+		attesterHistory := v.attesterHistoryByPubKey[pubKey]
+		attesterHistory = markAttestationForTargetEpoch(attesterHistory, indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
+		v.attesterHistoryByPubKey[pubKey] = attesterHistory
+		v.attesterHistoryByPubKeyLock.Unlock()
+	}
+
+	if featureconfig.Get().SlasherProtection && v.protector != nil {
+		if !v.protector.CommitAttestation(ctx, indexedAtt) {
+			if v.emitAccountMetrics {
+				ValidatorAttestFailVecSlasher.WithLabelValues(fmtKey).Inc()
+			}
+			return fmt.Errorf("made a slashable attestation, sourceEpoch: %dtargetEpoch: %d  "+
+				" found by external slasher service", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
+		}
+	}
+	return nil
+}
+
+// Given the validator public key, this gets the validator assignment.
+func (v *validator) duty(pubKey [48]byte) (*ethpb.DutiesResponse_Duty, error) {
+	if v.duties == nil {
+		return nil, errors.New("no duties for validators")
+	}
+
+	for _, duty := range v.duties.Duties {
+		if bytes.Equal(pubKey[:], duty.PublicKey) {
+			return duty, nil
+		}
+	}
+
+	return nil, fmt.Errorf("pubkey %#x not in duties", bytesutil.Trunc(pubKey[:]))
 }
 
 // Given validator's public key, this returns the signature of an attestation data.
@@ -262,61 +332,4 @@ func (v *validator) waitToSlotOneThird(ctx context.Context, slot uint64) {
 	startTime := slotutil.SlotStartTime(v.genesisTime, slot)
 	finalTime := startTime.Add(delay)
 	time.Sleep(roughtime.Until(finalTime))
-}
-
-func (v *validator) preSigningValidations(ctx context.Context, indexedAtt *ethpb.IndexedAttestation, pubKey [48]byte) error {
-	fmtKey := fmt.Sprintf("%#x", pubKey[:])
-	log := log.WithField("pubKey", fmt.Sprintf("%#x", bytesutil.Trunc(pubKey[:]))).WithField("slot", indexedAtt.Data.Slot)
-	if featureconfig.Get().ProtectAttester {
-		v.attesterHistoryByPubKeyLock.RLock()
-		attesterHistory := v.attesterHistoryByPubKey[pubKey]
-		v.attesterHistoryByPubKeyLock.RUnlock()
-		if isNewAttSlashable(attesterHistory, indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch) {
-			log.WithFields(logrus.Fields{
-				"sourceEpoch": indexedAtt.Data.Source.Epoch,
-				"targetEpoch": indexedAtt.Data.Target.Epoch,
-			}).Error("Attempted to make a slashable attestation, rejected")
-			if v.emitAccountMetrics {
-				metrics.ValidatorAttestFailVec.WithLabelValues(fmtKey).Inc()
-			}
-			return fmt.Errorf("sourceEpoch: %dtargetEpoch: %d  Attempted to make a slashable attestation,"+
-				" rejected by local slasher protection", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
-		}
-	}
-	if featureconfig.Get().SlasherProtection && v.protector != nil {
-		if !v.protector.VerifyAttestation(ctx, indexedAtt) {
-			log.WithFields(logrus.Fields{
-				"sourceEpoch": indexedAtt.Data.Source.Epoch,
-				"targetEpoch": indexedAtt.Data.Target.Epoch,
-			}).Error("Attempted to make a slashable attestation, rejected by external slasher service")
-			if v.emitAccountMetrics {
-				metrics.ValidatorAttestFailVecSlasher.WithLabelValues(fmtKey).Inc()
-			}
-			return fmt.Errorf("sourceEpoch: %dtargetEpoch: %d  Attempted to make a slashable attestation,"+
-				" rejected by external slasher service", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
-		}
-	}
-	return nil
-}
-
-func (v *validator) postSignatureUpdate(ctx context.Context, indexedAtt *ethpb.IndexedAttestation, pubKey [48]byte) error {
-	fmtKey := fmt.Sprintf("%#x", pubKey[:])
-	if featureconfig.Get().ProtectAttester {
-		v.attesterHistoryByPubKeyLock.Lock()
-		attesterHistory := v.attesterHistoryByPubKey[pubKey]
-		attesterHistory = markAttestationForTargetEpoch(attesterHistory, indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
-		v.attesterHistoryByPubKey[pubKey] = attesterHistory
-		v.attesterHistoryByPubKeyLock.Unlock()
-	}
-
-	if featureconfig.Get().SlasherProtection && v.protector != nil {
-		if !v.protector.CommitAttestation(ctx, indexedAtt) {
-			if v.emitAccountMetrics {
-				metrics.ValidatorAttestFailVecSlasher.WithLabelValues(fmtKey).Inc()
-			}
-			return fmt.Errorf("made a slashable attestation, sourceEpoch: %dtargetEpoch: %d  "+
-				" found by external slasher service", indexedAtt.Data.Source.Epoch, indexedAtt.Data.Target.Epoch)
-		}
-	}
-	return nil
 }
