+++ validator/client/attest_test.go
@@ -1,4 +1,4 @@
-package streaming
+package client
 
 import (
 	"context"
@@ -27,8 +27,7 @@ import (
 func TestRequestAttestation_ValidatorDutiesRequestFailure(t *testing.T) {
 	hook := logTest.NewGlobal()
 	validator, _, finish := setup(t)
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{}
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{}}
 	defer finish()
 
 	validator.SubmitAttestation(context.Background(), 30, validatorPubKey)
@@ -40,15 +39,13 @@ func TestAttestToBlockHead_SubmitAttestation_EmptyCommittee(t *testing.T) {
 
 	validator, _, finish := setup(t)
 	defer finish()
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 0,
 			Committee:      make([]uint64, 0),
 			ValidatorIndex: 0,
-		},
-	}
+		}}}
 	validator.SubmitAttestation(context.Background(), 0, validatorPubKey)
 	testutil.AssertLogsContain(t, hook, "Empty committee")
 }
@@ -58,15 +55,13 @@ func TestAttestToBlockHead_SubmitAttestation_RequestFailure(t *testing.T) {
 
 	validator, m, finish := setup(t)
 	defer finish()
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      make([]uint64, 111),
 			ValidatorIndex: 0,
-		},
-	}
+		}}}
 	m.validatorClient.EXPECT().GetAttestationData(
 		gomock.Any(), // ctx
 		gomock.AssignableToTypeOf(&ethpb.AttestationDataRequest{}),
@@ -99,15 +94,14 @@ func TestAttestToBlockHead_AttestsCorrectly(t *testing.T) {
 	hook := logTest.NewGlobal()
 	validatorIndex := uint64(7)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
 		},
-	}
+	}}
 
 	beaconBlockRoot := bytesutil.ToBytes32([]byte("A"))
 	targetRoot := bytesutil.ToBytes32([]byte("B"))
@@ -176,15 +170,14 @@ func TestAttestToBlockHead_BlocksDoubleAtt(t *testing.T) {
 	defer finish()
 	validatorIndex := uint64(7)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
 		},
-	}
+	}}
 	beaconBlockRoot := bytesutil.ToBytes32([]byte("A"))
 	targetRoot := bytesutil.ToBytes32([]byte("B"))
 	sourceRoot := bytesutil.ToBytes32([]byte("C"))
@@ -302,15 +295,14 @@ func TestAttestToBlockHead_BlocksSurroundAtt(t *testing.T) {
 	defer finish()
 	validatorIndex := uint64(7)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
 		},
-	}
+	}}
 	beaconBlockRoot := bytesutil.ToBytes32([]byte("A"))
 	targetRoot := bytesutil.ToBytes32([]byte("B"))
 	sourceRoot := bytesutil.ToBytes32([]byte("C"))
@@ -350,15 +342,14 @@ func TestAttestToBlockHead_BlocksSurroundedAtt(t *testing.T) {
 	defer finish()
 	validatorIndex := uint64(7)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
 		},
-	}
+	}}
 	beaconBlockRoot := bytesutil.ToBytes32([]byte("A"))
 	targetRoot := bytesutil.ToBytes32([]byte("B"))
 	sourceRoot := bytesutil.ToBytes32([]byte("C"))
@@ -434,15 +425,13 @@ func TestAttestToBlockHead_DoesAttestAfterDelay(t *testing.T) {
 	validator.genesisTime = uint64(roughtime.Now().Unix())
 	validatorIndex := uint64(5)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
-		},
-	}
+		}}}
 
 	m.validatorClient.EXPECT().GetAttestationData(
 		gomock.Any(), // ctx
@@ -473,15 +462,13 @@ func TestAttestToBlockHead_CorrectBitfieldLength(t *testing.T) {
 	defer finish()
 	validatorIndex := uint64(2)
 	committee := []uint64{0, 3, 4, 2, validatorIndex, 6, 8, 9, 10}
-	validator.dutiesByEpoch = make(map[uint64][]*ethpb.DutiesResponse_Duty)
-	validator.dutiesByEpoch[0] = []*ethpb.DutiesResponse_Duty{
+	validator.duties = &ethpb.DutiesResponse{Duties: []*ethpb.DutiesResponse_Duty{
 		{
 			PublicKey:      validatorKey.PublicKey.Marshal(),
 			CommitteeIndex: 5,
 			Committee:      committee,
 			ValidatorIndex: validatorIndex,
-		},
-	}
+		}}}
 	m.validatorClient.EXPECT().GetAttestationData(
 		gomock.Any(), // ctx
 		gomock.AssignableToTypeOf(&ethpb.AttestationDataRequest{}),
