+++ validator/client/log.go
@@ -1,4 +1,4 @@
-package polling
+package client
 
 import (
 	"fmt"
