+++ validator/client/metrics.go
@@ -1,14 +1,130 @@
-package streaming
+package client
 
 import (
 	"context"
 	"fmt"
 
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/prysmaticlabs/prysm/validator/client/metrics"
 	"github.com/sirupsen/logrus"
+)
+
+var (
+	// ValidatorStatusesGaugeVec used to track validator statuses by public key.
+	ValidatorStatusesGaugeVec = promauto.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Namespace: "validator",
+			Name:      "statuses",
+			Help:      "validator statuses: 0 UNKNOWN, 1 DEPOSITED, 2 PENDING, 3 ACTIVE, 4 EXITING, 5 SLASHING, 6 EXITED",
+		},
+		[]string{
+			// Validator pubkey.
+			"pubkey",
+		},
+	)
+	// ValidatorAggSuccessVec used to count successful aggregations.
+	ValidatorAggSuccessVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "successful_aggregations",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorAggFailVec used to count failed aggregations.
+	ValidatorAggFailVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "failed_aggregations",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorProposeSuccessVec used to count successful proposals.
+	ValidatorProposeSuccessVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "successful_proposals",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorProposeFailVec used to count failed proposals.
+	ValidatorProposeFailVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "failed_proposals",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorProposeFailVecSlasher used to count failed proposals by slashing protection.
+	ValidatorProposeFailVecSlasher = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "validator_proposals_rejected_total",
+			Help: "Count the block proposals rejected by slashing protection.",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorBalancesGaugeVec used to keep track of validator balances by public key.
+	ValidatorBalancesGaugeVec = promauto.NewGaugeVec(
+		prometheus.GaugeOpts{
+			Namespace: "validator",
+			Name:      "balance",
+			Help:      "current validator balance.",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorAttestSuccessVec used to count successful attestations.
+	ValidatorAttestSuccessVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "successful_attestations",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorAttestFailVec used to count failed attestations.
+	ValidatorAttestFailVec = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Namespace: "validator",
+			Name:      "failed_attestations",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
+	// ValidatorAttestFailVecSlasher used to count failed attestations by slashing protection.
+	ValidatorAttestFailVecSlasher = promauto.NewCounterVec(
+		prometheus.CounterOpts{
+			Name: "validator_attestations_rejected_total",
+			Help: "Count the attestations rejected by slashing protection.",
+		},
+		[]string{
+			// validator pubkey
+			"pubkey",
+		},
+	)
 )
 
 // LogValidatorGainsAndLosses logs important metrics related to this validator client's
@@ -41,7 +157,7 @@ func (v *validator) LogValidatorGainsAndLosses(ctx context.Context, slot uint64)
 	if v.emitAccountMetrics {
 		for _, missingPubKey := range resp.MissingValidators {
 			fmtKey := fmt.Sprintf("%#x", missingPubKey[:])
-			metrics.ValidatorBalancesGaugeVec.WithLabelValues(fmtKey).Set(0)
+			ValidatorBalancesGaugeVec.WithLabelValues(fmtKey).Set(0)
 		}
 	}
 
@@ -60,6 +176,7 @@ func (v *validator) LogValidatorGainsAndLosses(ctx context.Context, slot uint64)
 			v.prevBalance[pubKeyBytes] = params.BeaconConfig().MaxEffectiveBalance
 		}
 
+		fmtKey := fmt.Sprintf("%#x", pubKey)
 		truncatedKey := fmt.Sprintf("%#x", pubKey[:8])
 		if v.prevBalance[pubKeyBytes] > 0 {
 			newBalance := float64(resp.BalancesAfterEpochTransition[i]) / gweiPerEth
@@ -78,7 +195,7 @@ func (v *validator) LogValidatorGainsAndLosses(ctx context.Context, slot uint64)
 				"percentChange":        fmt.Sprintf("%.5f%%", percentNet*100),
 			}).Info("Previous epoch voting summary")
 			if v.emitAccountMetrics {
-				metrics.ValidatorBalancesGaugeVec.WithLabelValues(truncatedKey).Set(newBalance)
+				ValidatorBalancesGaugeVec.WithLabelValues(fmtKey).Set(newBalance)
 			}
 		}
