--- validator/client/metrics/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["metrics.go"],
-    importpath = "github.com/prysmaticlabs/prysm/validator/client/metrics",
-    visibility = ["//validator/client:__subpackages__"],
-    deps = [
-        "@com_github_prometheus_client_golang//prometheus:go_default_library",
-        "@com_github_prometheus_client_golang//prometheus/promauto:go_default_library",
-    ],
-)
