--- validator/client/polling/BUILD.bazel
@@ -1,98 +1,0 @@
-load("@prysm//tools/go:def.bzl", "go_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "aggregate.go",
-        "attest.go",
-        "log.go",
-        "metrics.go",
-        "propose.go",
-        "runner.go",
-        "service.go",
-        "validator.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/validator/client/polling",
-    visibility = ["//validator:__subpackages__"],
-    deps = [
-        "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/state/stateutil:go_default_library",
-        "//proto/slashing:go_default_library",
-        "//shared/blockutil:go_default_library",
-        "//shared/bls:go_default_library",
-        "//shared/bytesutil:go_default_library",
-        "//shared/featureconfig:go_default_library",
-        "//shared/grpcutils:go_default_library",
-        "//shared/hashutil:go_default_library",
-        "//shared/params:go_default_library",
-        "//shared/roughtime:go_default_library",
-        "//shared/slotutil:go_default_library",
-        "//validator/client/metrics:go_default_library",
-        "//validator/db:go_default_library",
-        "//validator/db/kv:go_default_library",
-        "//validator/keymanager/v1:go_default_library",
-        "//validator/slashing-protection:go_default_library",
-        "@com_github_dgraph_io_ristretto//:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-        "@com_github_gogo_protobuf//types:go_default_library",
-        "@com_github_grpc_ecosystem_go_grpc_middleware//:go_default_library",
-        "@com_github_grpc_ecosystem_go_grpc_middleware//retry:go_default_library",
-        "@com_github_grpc_ecosystem_go_grpc_middleware//tracing/opentracing:go_default_library",
-        "@com_github_grpc_ecosystem_go_grpc_prometheus//:go_default_library",
-        "@com_github_hashicorp_golang_lru//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
-        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@io_opencensus_go//plugin/ocgrpc:go_default_library",
-        "@io_opencensus_go//trace:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//codes:go_default_library",
-        "@org_golang_google_grpc//credentials:go_default_library",
-        "@org_golang_google_grpc//metadata:go_default_library",
-        "@org_golang_google_grpc//status:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    size = "small",
-    srcs = [
-        "aggregate_test.go",
-        "attest_test.go",
-        "fake_validator_test.go",
-        "propose_test.go",
-        "runner_test.go",
-        "service_test.go",
-        "validator_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//beacon-chain/core/helpers:go_default_library",
-        "//proto/slashing:go_default_library",
-        "//shared:go_default_library",
-        "//shared/bls:go_default_library",
-        "//shared/bytesutil:go_default_library",
-        "//shared/featureconfig:go_default_library",
-        "//shared/keystore:go_default_library",
-        "//shared/mock:go_default_library",
-        "//shared/params:go_default_library",
-        "//shared/roughtime:go_default_library",
-        "//shared/slotutil:go_default_library",
-        "//shared/testutil:go_default_library",
-        "//validator/accounts/v1:go_default_library",
-        "//validator/db/testing:go_default_library",
-        "//validator/keymanager/v1:go_default_library",
-        "//validator/testing:go_default_library",
-        "@com_github_gogo_protobuf//types:go_default_library",
-        "@com_github_golang_mock//gomock:go_default_library",
-        "@com_github_hashicorp_golang_lru//:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
-        "@in_gopkg_d4l3k_messagediff_v1//:go_default_library",
-    ],
-)
