+++ validator/node/BUILD.bazel
@@ -27,8 +27,7 @@ go_library(
         "//shared/prometheus:go_default_library",
         "//shared/tracing:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/client/polling:go_default_library",
-        "//validator/client/streaming:go_default_library",
+        "//validator/client:go_default_library",
         "//validator/db/kv:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/keymanager/v1:go_default_library",
